requirements.txt
aws-cdk-lib==2.150.0
constructs>=10.0.0,<11.0.0
boto3==1.34.144
requests==2.32.3
tenacity==8.5.0
aws-xray-sdk==2.14.0
